requirements.txt
asgiref==3.6.0
Django==4.2
django-crispy-forms==2.0
django-environ==0.10.0
django-quill-editor==0.1.40
gunicorn==20.1.0
Pillow==9.5.0
sqlparse==0.4.3
